Split up common functionality from Trainer (which will eventually become LSTMTrainer) so we can build an EnsembleTrainer
